+++ frontend/app/tests/e2e/pages/assets-manager-page.ts
@@ -6,24 +6,24 @@ export class AssetsManagerPage {
   }
 
   ignoredAssets() {
-    cy.get('[data-cy=asset-filter]').click();
+    cy.get('[data-cy=status-filter]').click();
     return cy
       .get('[data-cy=asset-filter-show_only]')
       .invoke('text')
       .then(text => {
-        cy.get('[data-cy=asset-filter]').click();
+        cy.get('[data-cy=status-filter]').click();
         cy.wrap(text.replace(/[^\d.]/g, ''));
       });
   }
 
   ignoredAssetCount(number: number) {
-    cy.get('[data-cy=asset-filter]').click();
+    cy.get('[data-cy=status-filter]').click();
     cy.get('[data-cy=asset-filter-menu]').should('be.visible');
     cy.get('[data-cy=asset-filter-show_only]').should(
       'include.text',
       number.toString()
     );
-    cy.get('[data-cy=asset-filter]').click();
+    cy.get('[data-cy=status-filter]').click();
     cy.get('[data-cy=asset-filter-menu]').should('not.be.visible');
   }
 
@@ -71,12 +71,12 @@ export class AssetsManagerPage {
   }
 
   selectShowAll(): void {
-    cy.get('[data-cy=asset-filter]').scrollIntoView();
-    cy.get('[data-cy=asset-filter]').click();
+    cy.get('[data-cy=status-filter]').scrollIntoView();
+    cy.get('[data-cy=status-filter]').click();
     cy.get('[data-cy=asset-filter-menu]').should('be.visible');
     cy.get('[data-cy=asset-filter-none]').scrollIntoView();
     cy.get('[data-cy=asset-filter-none]').click();
-    cy.get('[data-cy=asset-filter]').click();
+    cy.get('[data-cy=status-filter]').click();
     cy.get('[data-cy=asset-filter-menu]').should('not.be.visible');
   }