+++ rotkehlchen/exchanges/constants.py
@@ -9,7 +9,6 @@ SUPPORTED_EXCHANGES = (
     Location.BITMEX,
     Location.BITPANDA,
     Location.BITSTAMP,
-    Location.BITTREX,
     Location.COINBASE,
     Location.COINBASEPRO,
     Location.GEMINI,
@@ -22,6 +21,6 @@ SUPPORTED_EXCHANGES = (
     Location.WOO,
 )
 EXCHANGES_WITH_PASSPHRASE = (Location.COINBASEPRO, Location.KUCOIN, Location.OKX)
-DEAD_EXCHANGES = (Location.FTX, Location.FTXUS)
+DEAD_EXCHANGES = (Location.FTX, Location.FTXUS, Location.BITTREX)
 # Exchanges for which we allow import via CSV
 ALL_SUPPORTED_EXCHANGES = SUPPORTED_EXCHANGES + EXTERNAL_EXCHANGES + DEAD_EXCHANGES