+++ rotkehlchen/exchanges/utils.py
@@ -42,7 +42,7 @@ def deserialize_asset_movement_address(
     """Gets the address from an asset movement mapping making sure that if it's
     an ethereum deposit/withdrawal the address is returned checksummed"""
     value = get_key_if_has_val(mapping, key)
-    if value and asset == A_ETH:
+    if value and (asset == A_ETH or asset.is_evm_token()):
         try:
             value = to_checksum_address(value)
         except ValueError: